Cleanup

+++ include/madgwick_filter.h
@@ -17,7 +17,7 @@ public:
 
 private:
   // **** paramaters
-  double gain_ = 1.0;                                  // algorithm gain
+  float gain_ = 1.0;                                  // algorithm gain
   const String world_frame[3] = {"NWU", "ENU", "NED"}; // NWU, ENU, NED
   int world_frame_id;                                  // 0 - NWU, 1 - ENU, 2 - NED
 
@@ -31,7 +31,7 @@ private:
   long last_time_;
 
 public:
-  void setAlgorithmGain(double gain)
+  void setAlgorithmGain(float gain)
   {
     /*Gain of the filter.
      Higher values lead to faster convergence but more noise.
